Update pbl_compare.f90
Correction for #3

+++ src/pbl_compare.f90
@@ -18,8 +18,8 @@ module pbl_compare
     type CompareType
         logical                             :: lGo = .false.
         integer                             :: iNumData
-        real(8), dimension(:), allocatable  :: rvPrimary
-        real(8), dimension(:), allocatable  :: rvSecondary
+        real, dimension(:), allocatable     :: rvPrimary
+        real, dimension(:), allocatable     :: rvSecondary
     contains
         procedure   :: Set
         procedure   :: FB       ! Fractional bias
@@ -83,15 +83,15 @@ contains
     
         ! Routine arguments
         class(CompareType), intent(inout)   :: this
-        real(8)                             :: rFB
+        real                                :: rFB
         
         ! Locals
-        real(8) :: rNumerator
-        real(8) :: rDenominator
+        real    :: rNumerator
+        real    :: rDenominator
         
         ! Check execution may start
         if(.not.this % lGo) then
-            rFB = -9999.9d0
+            rFB = -9999.9
             return
         end if
